Moved Response from Models to Utils
