+++ app/core/app_manager.py
@@ -1,7 +1,7 @@
 from app.core.task_manager import TaskManager
 from app.core.data_manager import DataManager
 from app.core.device_manager import DeviceManager
-from ..models import Response
+from .utils.response import Response
 from .data.command import Command
 from .log import Log
 from .utils.errors import IdError
