+++ app/core/utils/response.py
@@ -1,0 +1,20 @@
+from flask import jsonify
+
+
+class Response:
+    def __init__(self, success, data, cause):
+        self.cause = cause
+        self.data = data
+        self.success = success
+
+    def __str__(self):
+        return str(self.__dict__)
+
+    def to_json(self):
+        return jsonify(
+            {
+                "success": self.success,
+                "cause": str(self.cause),
+                "data": self.data,
+            }
+        )
